Replaced size_t with uint32_t for now to make sure the trace files work both on x86 and x64

+++ core/file.c
@@ -20,7 +20,7 @@ bool file_exists(const char* path)
     }
 }
 
-bool file_path_indexed(char* path, size_t path_size, const char* dir, const char* name, const char* ext, uint32_t* index)
+bool file_path_indexed(char* path, uint32_t path_size, const char* dir, const char* name, const char* ext, uint32_t* index)
 {
     // only up to four decimals are allowed
     if (*index > 9999) {
