+++ retrace/retrace.h
@@ -3,4 +3,4 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-void plugin_set_rdram_size(size_t size);
+void plugin_set_rdram_size(uint32_t size);
